Search for a given magic sum
I originally planned to enumerate Pythagorean for all magic sums at the
same time, but we’d then have to store all their A, B, C values in
memory which isn’t really practical. We could try to offload to
individual files (e.g. named after the magic sum) but we’d then have to
keep a lot of file handles around and it generally seems problematic.

Instead, I think it’s better to repeatedly run the program for each
target sum. I expect the enumeration of Pythagorean triples to be
relative fast compared to the subgraph isomorphism checking that I’m
planning to add, so I think this approach is fine for now. It also means
we can short circuit the for loops once we’ve exceeded the target sum.