Skip symmetrical Pythagorean triples
Previously, we were enumerating symmetrical triples, e.g.

100000011 = 81234169 + 8281 + 18757561
100000011 = 81234169 + 18757561 + 8281

To avoid enumerating B + C as well as C + B we can break once C exceeds
B. That means that A, B, C will always be descending since A is always
the highest number that we have just enumerated in the outer for loop.