Add the magic square subgraph to pattern.vf
I wrote this by hand but I'm pretty confident it's correct. We only need to add
edges in one direction since we are using the -u flag for vf3. I've used to
convention of adding an add from the smaller number to the larger number. The
graph nodes are mapped onto a magic square as follows:

```
0 | 1 | 2
--|---|--
3 | 4 | 5
--|---|--
6 | 7 | 8
```