Add high-level orchestration code
Add a command-line parser and add support for multiple threads. The
--start and --stride arguments allow resuming from a specific point and
allows work to be distributed across multiple machines, e.g. by starting
at 3 on one machine and 75 on another, then striding by 144 on both.

The program logs which magic sums have been tested so that we can resume
in case of power failure or if we need to use the computer for something
else. We can also specify the number of threads, in case you want to use
the computer while the search is running.